// types.ts
export interface SleepLog {
  id: string;
  hours: number;
  timestamp: string;
}

export interface StepLog {
  id: string;
  steps: number;
  timestamp: string;
}